Modularize the db plugin. A single file was getting a tad large
